D2: 拆 doctor_exam_question_edit_dialogs.dart 的 3 个 Dialog
AudioSettingDialog / GenerateAudioFromTextDialog / HintRuleEditDialog 各自抽到独立文件，
原 1032 行的 doctor_exam_question_edit_dialogs.dart 删除，
doctor_exam_question_edit.dart 改为直接 import 实际用到的 2 个新文件
(AudioSettingDialog / HintRuleEditDialog；GenerateAudioFromTextDialog 仅被 audio_setting_dialog 内部依赖)。

零行为变更，flutter analyze 零新增告警。
